% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.xts.R
\name{aggregate.xts}
\alias{aggregate.xts}
\alias{aggregate.formula}
\alias{agg_mdh}
\alias{agg_monday}
\alias{agg_yhour}
\alias{agg_hour}
\alias{agg_year}
\alias{agg_yday}
\alias{agg_mday}
\alias{agg_mon}
\alias{agg_wday}
\alias{agg_wdayhour}
\alias{agg_by_period}
\title{Aggregate xts object}
\usage{
\method{aggregate}{xts}(x, by, FUN, ..., wsize = 0)

\method{aggregate}{formula}(formula, data, FUN, ..., subset, na.action = stats::na.omit)

agg_mdh(x, ...)

agg_monday(x, ...)

agg_yhour(x, ...)

agg_hour(x, ...)

agg_year(x, ...)

agg_yday(x, ...)

agg_mday(x, ...)

agg_mon(x, ...)

agg_wday(x, ...)

agg_wdayhour(x, ...)

agg_by_period(x, period, FUN, ...)
}
\arguments{
\item{x, data}{An \code{xts} object}

\item{by}{A character string speficying overlapping time indexing to apply
the \code{FUN}}

\item{FUN}{an argument of type \code{function}}

\item{...}{additional arguments for \code{FUN}}

\item{wsize}{Window size to to take into account}

\item{formula}{a formula, such as y ~ x or cbind(y1, y2) ~ x1 + x2, where
the y variables are numeric data to be split into groups
according to the grouping x variables (usually factors).}

\item{subset}{an optional vector specifying a subset of observations to
be used.}

\item{na.action}{a function which indicates what should happen when the
data contain NA values. The default is to ignore missing
values in the given variables.}

\item{period}{Length of period of xts to aggregate.}
}
\description{
Aggregate \code{xts} object by month, day, year etc...
}
\details{
Valid values for the argument \code{on} include:
"sec", "min", "hour", "mday", "mon", "year", "wday", "yday", "yhour", "isdst".
\code{agg.*} functions are wrapper functions for overlapping time periods
}
