% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alone.R
\name{alone}
\alias{alone}
\alias{alone.formula}
\alias{alone_index}
\alias{alone_index.formula}
\title{Alone Value Index}
\usage{
alone(x, ...)

\method{alone}{formula}(formula, data, ...)

alone_index(x, ..., flat = FALSE)

\method{alone_index}{formula}(formula, data, ...)
}
\arguments{
\item{x}{An R object (\code{vector}, \code{matrix} or \code{data.frame})}

\item{...}{Unused.}

\item{formula}{a formula, such as y ~ x or cbind(y1, y2) ~ x1 + x2, where
the y variables are numeric data to be split into groups
according to the grouping x variables (usually factors).}

\item{data}{A \code{data.frame}}

\item{flat}{If \code{x} is a matrix-like object, it's interpreted as vector
and indices are returned according to this.}
}
\description{
This function returns values/indices of values between at least two
\code{NA}'s.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: S3 method for \code{formula} interface

\item \code{formula}: S3 method for \code{formula} interface
}}

