% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsarray.R
\name{as.tsarray}
\alias{as.tsarray}
\alias{as.tsarray.ts}
\alias{as.tsarray.zoo}
\title{Convert time-series object to a \code{tsarray} object}
\usage{
as.tsarray(x, ...)

\method{as.tsarray}{ts}(x, ...)

\method{as.tsarray}{zoo}(x, ...)
}
\arguments{
\item{x}{a \link{ts}, \code{\link[zoo:zoo]{zoo}}  or \link{xts} object}

\item{f}{Frequency (an integer value) or convert to matrix criteria for
the object. Possible values are \code{(\%Y, \%m, \%d, \%H, \%M)}
for year, month, day, hour and minute, respectively.}
}
\value{
A \code{[tsarray]} object
}
\description{
Convert time-series object to a \code{tsarray} object by longest frequency of
time-series object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ts}: S3 method for ts object

\item \code{zoo}: S3 method for zoo object
}}

\keyword{internal}
