% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\alias{boxcox.lambda}
\alias{boxcox.plus}
\title{BoxCox Transformation}
\usage{
boxcox(x, lambda = c(-2, 2), inverse = FALSE, log_10 = FALSE)

boxcox.lambda(x, lambda = c(-2, 2), log_10 = FALSE)

boxcox.plus(x, plus = c(0, 100), log_10 = FALSE)
}
\arguments{
\item{x}{A numeric vector-like R object.}

\item{lambda}{If \code{lambda} is a single value, transformation is
accomplished using this value. If a 2-length vector, best
lambda value is calculated for transformation. \code{lambda}
always must be 2-length vector for \code{boxcox.lambda}
function.}

\item{inverse}{If \code{TRUE}, inverse boxcox transformation is calculated.}

\item{log_10}{If \code{TRUE}, transformation is accomplished by \code{log10}
function; otherwise, natural \code{log} function is used.}
}
\description{
\code{boxcox} function applies boxcox-transformation to \code{x}. Calculated
best \code{lambda} value for the transformation is attached to \code{x} as
attribute. (See: \code{attr(x, "lambda")}). If \code{x} contains any value
equal or less than \code{0}, it will be set to \code{NA} for the sake of
transformation.

\code{boxcox.lambda} function calculates best lambda value for
             boxcox transformation in the interval of \code{lambda} values.

\code{boxcox.plus} calculates a value to add \code{x} values
             which maximizes \code{loglik} function for \code{log} or
             \code{log10} transformation.
}
