% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_object_colnames.R
\name{check_object_colnames}
\alias{check_object_colnames}
\title{Check object colnames}
\usage{
check_object_colnames(object)
}
\description{
If column names of the object is not set, column names of the \code{object}
are set by \code{object} names. If the \code{object} does not have dimension,
it's dimension set as 1 column.
}
\keyword{internal}
