% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_formula.R
\name{eval_formula}
\alias{eval_formula}
\title{Evaluate formula on data}
\usage{
eval_formula(formula, data, ...)
}
\arguments{
\item{formula}{Formula to be evaluated}

\item{data}{\code{data.frame} for the formula}

\item{...}{Other params to be passed to \code{model.frame}}
}
\value{
Evaluated formula result
}
\description{
This function evaluates formula on data and returns evaluated data.
}
