% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.missing.R
\name{fill.missing}
\alias{fill.missing}
\alias{fill.missing.xts}
\alias{fill.missing.zoo}
\title{Fill Missing Data}
\usage{
fill.missing(x, fill = NA, inrange = range(x), ...)

\method{fill.missing}{xts}(x, fill = NA, inrange = range(zoo::index(x)), ...)

\method{fill.missing}{zoo}(x, ...)
}
\arguments{
\item{x}{An xts object.}

\item{fill}{see \link[zoo:na.fill]{zoo::na.fill}.}

\item{inrange}{Range of indices to be filled with \code{NA}.}

\item{...}{Additional arguments to be passed to \link[zoo:na.fill]{zoo::na.fill}.}
}
\value{
filled object
}
\description{
Fill missing time steps in an \link{xts} series and complete \code{NA} values
if requested.
}
