% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.zoo.R
\name{fortify.zoo}
\alias{fortify.zoo}
\title{Convenience Function for Plotting xts or zoo Objects with ggplot2}
\usage{
\method{fortify}{zoo}(
  model,
  data,
  names = c("Index", "Series", "Value"),
  melt = FALSE,
  sep = NULL,
  by = NULL,
  reduce = FALSE,
  reformat.Index = TRUE,
  ...
)
}
\arguments{
\item{model}{An object of class \code{zoo} to be converted to a
\code{data.frame}.}

\item{data, ...}{Not used.}

\item{names}{(list of) character vector(s). New names given to index/time
column, series indicator (if melted), and value column
(if melted). If only a subset of characters should be changed,
either NAs can be used or a named vector.}

\item{melt}{If \code{TRUE}, resulting \code{data.frame} is in long format.}

\item{sep}{If specified then the Series column is split into multiple
columns using sep as the split character.}

\item{by}{bla bla bla}

\item{reduce}{If \code{TRUE}, temporal resolution of the object is reduced.
This is useful especially for large \code{xts} objects.
See \code{\link[=reduce]{reduce()}}.}
}
\description{
\code{fortify.zoo} takes a zoo object and converts it into a data frame
(intended for ggplot2).
}
\keyword{internal}
