% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexx.R
\name{indexx}
\alias{indexx}
\title{indexx function}
\usage{
indexx(x, unit)
}
\arguments{
\item{x}{\code{POSIXt}, \code{xts} or \code{zoo} object.}

\item{unit}{Unit as character vector. Possible values are \code{sec, min,
hour, mday, mon, year, wday, yday, isdst} and \code{yearmon, wdayhour, md,
mdh, yhour, index}.}
}
\description{
Extreact various date properties from \code{POSIXt}, \code{xts} or \code{zoo}
objects.
}
\keyword{internal}
