% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_for_each.R
\name{lapply_for_each}
\alias{lapply_for_each}
\alias{lapply_for_each_month}
\alias{lapply_for_each_part}
\title{Apply function for each same months in a zoo series}
\usage{
lapply_for_each_month(x, FUN, ...)

lapply_for_each_part(x, FUN, ..., npart = 52)
}
\arguments{
\item{x}{A \code{zoo} series}

\item{FUN}{a function to apply.}

\item{...}{Other arguments passed to function.}

\item{npart}{Number of parts to split the data.}
}
\description{
Apply function for each same months in a zoo series
}
