% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_at.R
\name{mean_at}
\alias{mean_at}
\alias{mean_at.xts}
\title{Calculate mean at a time step}
\usage{
mean_at(x, at, delta = 2, fill = FALSE)

\method{mean_at}{xts}(x, at, delta = as.difftime(2, units = "hours"), fill = FALSE)
}
\arguments{
\item{x}{An xts object}

\item{at}{Dates that means will be calculated. Can be \link{POSIXt} or
\link{character} vector.}

\item{delta}{Time window to the left and right of \code{at} parameter.
for instance, if \code{delta = 2}, measurements at
22, 23, 00, 01 and 02 is used to calculate the mean at 00.}

\item{fill}{If \code{TRUE}, missing values are filled with \code{NA}.}
}
\value{
Mean values at \code{at}
}
\description{
Calculate means on specific time steps in a \code{delta} time window. \code{at}
must be in the interval of max and min dates of \code{x}.
}
\section{Methods (by class)}{
\itemize{
\item \code{mean_at(xts)}: Calculate means on specific dates in a \code{delta} time.

}}
\examples{
time_stamps <- seq(from = as.POSIXct("2024-01-01 00:00:00"),
                   to = as.POSIXct("2024-01-02 23:00:00"), by = "hour")
at <- seq(from = min(time_stamps), to = max(time_stamps), by = "12 hours")
hourly_data <- xts::xts(seq_along(time_stamps), order.by = time_stamps,
                        dimnames = list(NULL, "x"))
mean_at(hourly_data, at = at)
# missing observations
mean_at(hourly_data[-(13:14),], at = as.POSIXct("2024-01-01 12:00:00"), fill = TRUE)
}
