% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_on.R
\name{mean_on}
\alias{mean_on}
\alias{mean_on.xts}
\alias{mean_on.data.frame}
\title{Calculate Mean on periods}
\usage{
mean_on(x, on, trim, na.rm, ...)

\method{mean_on}{xts}(x, on, trim = 0, na.rm = FALSE, ...)

\method{mean_on}{data.frame}(x, on, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{an \code{xts} or \code{data.frame} object.}

\item{on}{The periods to find as character string}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from
each end of x before the mean is computed. (see \link{mean})}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds or a value between
(0, 100) to decide percent of NA values.}

\item{...}{Not used.}
}
\description{
Calculate Mean on periods
}
\section{Methods (by class)}{
\itemize{
\item \code{mean_on(xts)}: S3 method for xts

\item \code{mean_on(data.frame)}: S3 method for data.frame

}}
\examples{
time_stamps <- seq(from = as.POSIXct("2024-01-01 00:00:00"),
                   to = as.POSIXct("2024-01-31 23:00:00"), by = "hour")
hourly_data <- xts::xts(seq_along(time_stamps), order.by = time_stamps,
                        dimnames = list(NULL, "x"))
mean_on(hourly_data, on = "days")
}
