% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medpolish.R
\name{medpol}
\alias{medpol}
\alias{medpol.xts}
\alias{na.medpol}
\alias{na.medpol.matrix}
\alias{na.medpol.xts}
\title{Median Polish for xts time series}
\usage{
medpol(x, ...)

\method{medpol}{xts}(x, ...)

na.medpol(x, ...)

\method{na.medpol}{matrix}(x, ...)

\method{na.medpol}{xts}(x, ...)
}
\arguments{
\item{x}{A numeric matrix or xts object}

\item{...}{Other params passed to \code{\link[stats]{medpolish}}}
}
\description{
Fits an additive model (twoway decomposition) using Tukey's median
polish procedure. If xts object contains data belong to February 29, it's
excluded before medpolish and included after medpolish.

\code{na.medpolish} fills NA's by medpolish method.
}
\section{Methods (by class)}{
\itemize{
\item \code{xts}: S3 method for xts object

\item \code{matrix}: S3 method for matrix

\item \code{xts}: S3 method for xts object
}}

\seealso{
\code{\link[stats]{medpolish}}
}
