% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ae}
\alias{ae}
\alias{mae}
\alias{se}
\alias{mse}
\alias{rmse}
\title{Error Metrics}
\usage{
ae(actual, predicted = 0)

mae(actual, predicted = 0)

se(actual, predicted = 0)

mse(actual, predicted = 0)

rmse(actual, predicted = 0)
}
\arguments{
\item{actual}{Actual data}

\item{predicted}{Predicted data}
}
\description{
Error Metrics
}
