% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode2.R
\name{mode2}
\alias{mode2}
\title{Mode of Numeric vector}
\usage{
mode2(x)
}
\arguments{
\item{x}{A vector-like R object.}
}
\description{
If \code{x} is a double-precision numeric vector, function calculates peak
(maximum) of Kernel Density Estimation of \code{x} values. If \code{x} is
integer or another R object, it's calcualted highest number of occurrences
in \code{x}.
}
