% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstl2.R
\name{mstl2}
\alias{mstl2}
\alias{plot.mstl}
\title{Multiple seasonal decomposition}
\usage{
mstl2(
  x,
  lambda = NULL,
  s.window = 13,
  ...,
  iterate = 2,
  type = c("additive", "multiplicative")
)

\method{plot}{mstl}(
  x,
  labels = colnames(X),
  set.pars = list(mar = c(0, 6, 0, 6), oma = c(6, 0, 4, 0), tck = -0.01, mfrow = c(nplot,
    1)),
  main = NULL,
  range.bars = TRUE,
  ...,
  col.range = "light gray"
)
}
\arguments{
\item{x}{Univariate time series of class \code{msts} or \code{ts}.}

\item{lambda}{BoxCox transformation coefficient. If \code{"auto"}, lambda is
automatically selected. If \code{NULL} no transformation is done.}

\item{s.window}{Seasonal windows to be used in the decomposition. If scalar,
the same value is used for all seasonal components. Otherwise, it should be a vector
of the same length as the number of seasonal components.}

\item{...}{Other arguments are passed to \code{\link[stats:stl]{stats::stl()}}.}

\item{iterate}{Number of iterations to use to refine the seasonal component.}

\item{type}{Type of decomposition. \code{additive} or \code{multiplicative}.}

\item{labels}{character of length 4 giving the names of the component
time-series.}

\item{set.pars}{settings for par(.) when setting up the plot.}

\item{main}{plot main title.}

\item{range.bars}{logical indicating if each plot should have a bar at its
right side which are of equal heights in user coordinates.}

\item{col.range}{Colour to be used for the range bars, if plotted. Note this
appears after ... and so cannot be abbreviated.}
}
\description{
Decompose a time series into seasonal, trend and remainder components.
\code{robust} is \code{TRUE} as different than \code{\link[forecast:mstl]{forecast::mstl()}}.
Also BoxCox transformation is accomplished by
\code{\link[=boxcox]{boxcox()}} function.
}
\seealso{
\code{\link[stats:stl]{stats::stl()}}, \code{\link[forecast:mstl]{forecast::mstl()}}
}
