% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query.POSIXt}
\alias{query.zoo}
\title{Subset POSIXt/zoo/xts objects}
\usage{
query(x, query, which.i, ...)

\method{query}{POSIXt}(x, query, which.i = FALSE, ...)

\method{query}{zoo}(x, query, which.i = FALSE, ...)
}
\arguments{
\item{x}{A \code{POSIXt}, \code{zoo} and \code{xts} objects}

\item{query}{Condition to query object}

\item{which.i}{If \code{TRUE}, indices are returned}

\item{...}{Unused}
}
\description{
Return subsets of \code{POSIXt}, \code{zoo} and \code{xts} objects
which meet \code{query} condition.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXt}: S3 method for class \code{POSIXt}

\item \code{zoo}: S3 method for class \code{zoo}
}}

\examples{
library(xts)
data(sample_matrix)
x <- as.xts(sample_matrix, dateFormat="Date")
query(x, "month == 10")
}
\seealso{
\link[base]{subset}
}
