% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_stats.R
\name{seasonal_stats}
\alias{seasonal_stats}
\alias{seasonal_stats.xts}
\title{Seasonal Analysis}
\usage{
seasonal_stats(x, ...)

\method{seasonal_stats}{xts}(x, FUNS = c("mean_cl_boot", "median_quantile"), ...)
}
\arguments{
\item{x}{An hourly \code{xts} object}

\item{...}{Unused for now.}

\item{FUNS}{Functions to be applied over \code{x}}
}
\description{
Calculate daily, weekly, annual means over hours and long term means over days.
}
\section{Methods (by class)}{
\itemize{
\item \code{seasonal_stats(xts)}: S3 Method for \code{xts}

}}
