% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_array.R
\name{slice_array}
\alias{slice_array}
\title{Slice Array}
\usage{
slice_array(x, margin, value, drop = TRUE)
}
\arguments{
\item{x}{An \code{array}}

\item{margin}{Margin for slicing.}

\item{value}{Margin value for slicing.}

\item{drop}{Drop results?}
}
\value{
Numeric sliced object.
}
\description{
Slice array by \code{margin} and margin \code{value} .
}
\keyword{internal}
