% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_stats.R
\name{mean_cl_normal}
\alias{mean_cl_normal}
\alias{mean_cl_boot}
\alias{median_quantile}
\title{Summary statistics on a vector}
\usage{
mean_cl_normal(x, ...)

mean_cl_boot(x, B = 10, ...)

median_quantile(x, q = c(0.25, 0.75), na.rm = TRUE)
}
\arguments{
\item{x}{A numveric vector.}

\item{...}{Other parameters passed to the functions.}

\item{B}{number of bootstrap resamples for \code{mean_cl_boot}.}

\item{q}{2 quantiles to calculate.}

\item{na.rm}{Do you want to remove \code{NA}'s? Default is \code{TRUE}.}
}
\description{
Summary statistics on a vector
}
