% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{trend}
\alias{trend}
\alias{trend.zoo}
\alias{trend.data.frame}
\alias{detrend}
\title{(de)trend xts object}
\usage{
trend(x, ...)

\method{trend}{zoo}(x, ..., fun = c("loess", "lm", "med"), predict = TRUE)

\method{trend}{data.frame}(x, ..., predict = TRUE)

detrend(
  x,
  fun = c("loess", "lm", "med"),
  ...,
  type = c("additive", "multiplicative")
)
}
\arguments{
\item{x}{an xts object}

\item{...}{other parameters passed to \code{loess} or \code{lm} methods.}

\item{fun}{\code{"loess"} or \code{"lm"} string to choose the model.
Can be abbreviated up to first two characters.}

\item{predict}{If \code{TRUE} predicted results are returned.}

\item{type}{De-trending type. \code{additive} or \code{multiplicative}.}
}
\value{
For \code{trend} method, if \code{predict} is \code{FALSE}, model
object is returned; otherwise predicted values return. For
\code{detrend} method, trend is removed from \code{x} object and
de-trended object with same class is returned.
}
\description{
\code{trend} method creates model object for \code{loess} or \code{lm}
methods if \code{predict} is \code{FALSE}; otherwise predicted values are
calculated.
\code{detrend} method removes long term trend from \code{xts} or
\code{data.frame} object by \code{loess} or \code{lm} methods.
Multiple series are accepted and trends are calculated for each column of
objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{trend(zoo)}: S3 method for \code{zoo} object

\item \code{trend(data.frame)}: S3 method for \code{data.frame} object

}}
