% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtstools.R
\docType{data}
\name{weekday.name}
\alias{weekday.name}
\alias{weekday.abb}
\alias{ndays}
\alias{monthbreak}
\title{Time-series variables}
\format{
\code{weekday.name}: Name of weekdays

\code{weekday.name}: Short name of weekdays

\code{ndays}: Number of days in months
}
\usage{
weekday.name

weekday.abb

ndays

monthbreak(mult = 24)
}
\arguments{
\item{mult}{Multiplier for month break calculation}
}
\description{
Some general purpose variables which might be useful for time-series
manipulations.

\code{monthbreak} function calculates start of each month.
}
\keyword{datasets}
