% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xapply.R
\name{xapply}
\alias{xapply}
\alias{xapply.default}
\alias{xapply.zoo}
\title{Extended - Apply a Function over array Margins}
\usage{
xapply(x, ...)

\method{xapply}{default}(
  x,
  MARGIN,
  FUN,
  ...,
  step_size = 1,
  wsize = 1,
  clamped = TRUE,
  multiple = FALSE
)

\method{xapply}{zoo}(x, FUN, ..., step_size = 1, wsize = 1, clamped = TRUE, multiple = FALSE)
}
\arguments{
\item{...}{Other arguments passed to \code{FUN}.}

\item{step_size}{Step size of moving window over margin. This value can be a
vector. Each value in the vector represents step size for a
new window in previous window.}

\item{wsize}{Window size for each \code{step_size}. This value can be a
vector. Each value in the vector represents window size for a
new window in previous window.}

\item{clamped}{This parameter defines hoe the windows will behave at the
edges of margin. If \code{TRUE}, windows are extended through
end and begining of the margins.}

\item{multiple}{This argument determines how \code{xapply} function will
behave If \code{x} is an array having a dimension equal to
3. If \code{multiple = FALSE} and dimension is 3, matrices in
each third dimension are estimated as seperate matrices.
Default is \code{FALSE}.}

\item{An}{array, including matrix.}

\item{margin}{A vector giving the subscripts which the function will be
applied over. E.g., for a matrix 1 indicates rows, 2 indicates
columns, c(1, 2) indicates rows and columns. Where X has named
dimnames, it can be a character vector selecting dimension
names.}
}
\description{
Returns a vector or array or list of values obtained by applying a function
to margins of an array or matrix.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default S3 method

\item \code{zoo}: S3 method for zoo object
}}

