% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtstools.R
\docType{package}
\name{xtstools-package}
\alias{xtstools}
\alias{xtstools-package}
\title{xtstools: Useful time-series tools}
\description{
scalex provides functions for climatological standardization
}
\author{
İsmail SEZEN \href{mailto:sezenismail@gmail.com}{sezenismail@gmail.com}
}
