% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\alias{trim.default}
\alias{trim.zoo}
\title{Trim data}
\usage{
trim(x, trim, na.rm)

\method{trim}{default}(x, trim = 0.15, na.rm = FALSE)

\method{trim}{zoo}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{trim}{Fraction of observations to be trimmed from each end of \code{x}.
Number of trimmed values are calculated on Non-NA values.}
}
\description{
Trim the data by sorting from each end. If \code{x} has dimension more than
one, each column trimmed sperately.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default S3 method

\item \code{zoo}: S3 method for class \code{zoo}
}}

